pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
graphene-django-cud = "^0.11.0"
django-graphql-jwt = "^0.4.0"
django-mapbox-location-field = "^2.0.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
pytest = "^7.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
